Move TryDeserialize<>.From<> into tests.
