+++ src/ExRam.Gremlinq.Core/Extensions/GremlinQueryFragmentDeserializerClassExtensions.cs
@@ -1,26 +1,10 @@
 ﻿using System.Diagnostics.CodeAnalysis;
-
 using ExRam.Gremlinq.Core.Deserialization;
 
 namespace ExRam.Gremlinq.Core
 {
     public static class GremlinQueryFragmentDeserializerClassExtensions
     {
-        public readonly struct FluentForClass<TRequested>
-            where TRequested : class
-        {
-            private readonly IGremlinQueryFragmentDeserializer _deserializer;
-
-            public FluentForClass(IGremlinQueryFragmentDeserializer deserializer)
-            {
-                _deserializer = deserializer;
-            }
-
-            public TRequested? From<TSerialized>(TSerialized serialized, IGremlinQueryEnvironment environment) => _deserializer.TryDeserialize<TSerialized, TRequested>(serialized, environment, out var value)
-                ? value
-                : default;
-        }
-
         private sealed class FixedTypeDeserializationTransformationFactory<TStaticSerialized, TStaticRequested> : IDeserializationTransformationFactory
            where TStaticRequested : class
         {
@@ -61,12 +45,6 @@ namespace ExRam.Gremlinq.Core
                     ? (IDeserializationTransformation<TSerialized, TRequested>)(object)new FixedTypeDeserializationTransformation<TSerialized>(_func)
                     : null;
             }
-        }
-
-        public static FluentForClass<TRequested> TryDeserialize<TRequested>(this IGremlinQueryFragmentDeserializer deserializer)
-            where TRequested : class
-        {
-            return new FluentForClass<TRequested>(deserializer);
         }
 
         public static IGremlinQueryFragmentDeserializer Override<TSerialized, TRequested>(this IGremlinQueryFragmentDeserializer fragmentDeserializer, Func<TSerialized, IGremlinQueryEnvironment, IGremlinQueryFragmentDeserializer, TRequested?> func)
