+++ src/ExRam.Gremlinq.Core/Extensions/GremlinQueryFragmentDeserializerStructExtensions.cs
@@ -5,21 +5,6 @@ namespace ExRam.Gremlinq.Core
 {
     public static class GremlinQueryFragmentDeserializerStructExtensions
     {
-        public readonly struct FluentForStruct<TRequested>
-            where TRequested : struct
-        {
-            private readonly IGremlinQueryFragmentDeserializer _deserializer;
-
-            public FluentForStruct(IGremlinQueryFragmentDeserializer deserializer)
-            {
-                _deserializer = deserializer;
-            }
-
-            public TRequested? From<TSerialized>(TSerialized serialized, IGremlinQueryEnvironment environment) => _deserializer.TryDeserialize<TSerialized, TRequested>(serialized, environment, out var value)
-                ? value
-                : default(TRequested?);
-        }
-
         private sealed class FixedTypeDeserializationTransformationFactory<TStaticSerialized, TStaticRequested> : IDeserializationTransformationFactory
             where TStaticRequested : struct
         {
@@ -60,12 +45,6 @@ namespace ExRam.Gremlinq.Core
                     ? (IDeserializationTransformation<TSerialized, TRequested>)(object)new FixedTypeDeserializationTransformation<TSerialized>(_func)
                     : null;
             }
-        }
-
-        public static FluentForStruct<TRequested> TryDeserialize<TRequested>(this IGremlinQueryFragmentDeserializer deserializer)
-            where TRequested : struct
-        {
-            return new FluentForStruct<TRequested>(deserializer);
         }
 
         public static IGremlinQueryFragmentDeserializer Override<TSerialized, TRequested>(this IGremlinQueryFragmentDeserializer fragmentDeserializer, Func<TSerialized, IGremlinQueryEnvironment, IGremlinQueryFragmentDeserializer, TRequested?> func)
